backend/requirements.txt
aiofiles==23.1.0
anyio==4.9.0
assemblyai==0.37.0
backoff==2.2.1
bcrypt==4.0.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.1.1
exceptiongroup==1.3.0
fastapi==0.95.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
loguru==0.7.0
packaging==25.0
passlib==1.7.4
pyasn1==0.6.1
pydantic==1.10.22
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.30.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.12
psycopg2-binary==2.9.9
starlette==0.26.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.22.0
websockets==15.0.1
